requirements.txt
python-telegram-bot==13.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
requests==2.31.0
flask==3.0.0
